+++ .eslintrc.js
@@ -2,22 +2,35 @@ module.exports = {
   env: {
     browser: true,
     es6: true,
+    jest: true,
     node: true
   },
   extends: [
-    'eslint:recommended',
+    'plugin:@typescript-eslint/recommended',
+    'prettier/@typescript-eslint',
     'plugin:prettier/recommended'
   ],
   parserOptions: {
-    ecmaVersion: 2019,
+    ecmaVersion: 2020,
     sourceType: 'module'
   },
-  plugins: [],
+  parser: '@typescript-eslint/parser',
+  plugins: ['@typescript-eslint'],
   rules: {
-    'indent': ['error', 2, { 'SwitchCase': 1 }],
+    '@typescript-eslint/no-unused-vars': 'error',
     'no-trailing-spaces': 'error',
     'object-curly-spacing': ['error', 'always'],
-    'quotes': ['error', 'single', { 'allowTemplateLiterals': true }],
-    'semi': ['error', 'never']
-  }
+    quotes: ['error', 'single', { allowTemplateLiterals: true }],
+    semi: ['error', 'never']
+  },
+  settings: {},
+  overrides: [
+    {
+      files: ['*.js'],
+      rules: {
+        '@typescript-eslint/explicit-function-return-type': 'off',
+        '@typescript-eslint/no-var-requires': 'off'
+      }
+    }
+  ]
 }