+++ source/hexToRgbaCss.ts
@@ -1,8 +1,10 @@
+import { numberAlpha, stringHex, stringRgba } from './customTypes'
 import hexToRgba from './hexToRgba'
 import rgbaToRgbaCss from './rgbaToRgbaCss'
 
-// hexToRgbaCss :: (String, Number) -> String Rgba
-const hexToRgbaCss = (str, alpha) => {
+type HexToRgbaCss = (hex: stringHex, alpha?: numberAlpha) => stringRgba
+
+const hexToRgbaCss: HexToRgbaCss = (str, alpha) => {
   const [r, g, b, _a] = hexToRgba(str)
   const a = alpha || _a
   return rgbaToRgbaCss([r, g, b, a])