+++ source/rgbaCssToHex.js
@@ -1,9 +1,0 @@
-import parseRgba from './parseRgba'
-import rgbaToHex from './rgbaToHex'
-
-// rgbaCssToHex :: String Rgba -> String Hex
-const rgbaCssToHex = x => {
-  return rgbaToHex(parseRgba(x))
-}
-
-export default rgbaCssToHex