+++ CHANGELOG.md
@@ -4,6 +4,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.0] - 2020-03-15
+
+### Added
+* typescript support
+* other bundles (minified CJS, minified UMD)
+
+### Changed
+* license from ISC to BSD-3
+
+### Fixed
+* security vulnerabilities from dependencies
+
+### Removed
+* `browser` entry point
+
 ## [0.3.1] - 2019-12-29
 
 ### Changed